// src/api/generated/ActivityEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ActivityDirection } from "./ActivityDirection";
import type { ActivityStatus } from "./ActivityStatus";

export type ActivityEntry = { txid: string, direction: ActivityDirection, status: ActivityStatus, amount_atomic: string, fee_atomic: string | null, height: string | null, timestamp: string | null, payment_id: string | null, };